Change command handler to use defined size int

+++ raat/devices/ENC28J60.device-plugin/ENC28J60RAAT.cpp
@@ -143,7 +143,7 @@ int ENC28J60RAAT::handle_set_command(char const * const command, char * reply)
     return reply_length;
 }
 
-int ENC28J60RAAT::command_handler(char const * const command, char * reply)
+uint16_t ENC28J60RAAT::command_handler(char const * const command, char * reply)
 {
     int reply_length = 0;
