+++ raat/devices/MT8870.device-plugin/MT8870.cpp
@@ -111,7 +111,7 @@ void MT8870::setup()
     this->reset();
 }
 
-int MT8870::command_handler(char const * const command, char * reply)
+uint16_t MT8870::command_handler(char const * const command, char * reply)
 {
     int reply_length = 0;
     if (command[0] == 'R')
