+++ raat/devices/adafruit-neopixel.device-plugin/adafruit-neopixel-raat.hpp
@@ -12,7 +12,7 @@ public:
     void setup();
     void reset();
     void tick();
-    int command_handler(char const * const command, char * reply);
+    uint16_t command_handler(char const * const command, char * reply);
     Adafruit_NeoPixel& pixels();
     void set_pixels(uint8_t range_min, uint8_t range_max, EightBitRGBParam& param);
     void set_pixels(uint8_t range_min, uint8_t range_max, uint8_t r, uint8_t g, uint8_t b);
