+++ raat/devices/potential-divider.device-plugin/potential-divider.hpp
@@ -15,7 +15,7 @@ public:
     void setup();
     void reset();
     void tick();
-    int command_handler(char const * const command, char * reply);
+    uint16_t command_handler(char const * const command, char * reply);
     
     float resistance();
 private:
